ALA-Web/backend/requirements.txt
fastapi
uvicorn
python-multipart
# Reuse existing dependencies
torch
transformers
timm
einops
opencv-python
Pillow
numpy
supervision
autodistill
autodistill-grounded-sam-2
autodistill-florence-2
autodistill-yolov8
roboflow
